# api/openapi.yaml
openapi: 3.1.0
info:
  title: Crabit API
  version: 0.0.1
  description: |
    Owner-facing Card Balance Account and Wish operations plus academy-facing
    Shared Card reads. Resource ownership and current social visibility are
    hidden with resource-specific 404 responses. All timestamps are RFC 3339
    UTC Z values, all KRW values are bounded integer won, and all error bodies
    use the common ErrorEnvelope.
servers:
  - url: /
...
paths:
...
  /v1/card-balance-accounts/{cardBalanceAccountId}/wishes/{wishId}/deposits:
    parameters:
      - $ref: '#/components/parameters/CardBalanceAccountId'
      - $ref: '#/components/parameters/WishId'
    post:
      tags: [Wishes]
      operationId: depositToWish
      summary: Deposit Card Balance Account funds into one Wish
      description: Performs PRE_DEPOSIT lookup internally. Provider failure leaves the Wish unchanged; a persisted mismatch observation locks and rejects only this deposit.
      security: [{SeedBearer: []}]
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/WishAmountCommand'}
      responses:
        '200': {$ref: '#/components/responses/WishMutationSuccess'}
        '400': {$ref: '#/components/responses/MalformedOrIdempotencyRequired'}
        '401': {$ref: '#/components/responses/AuthRequired'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/WishOrAccountNotFound'}
        '409': {$ref: '#/components/responses/DepositConflict'}
        '422': {$ref: '#/components/responses/InvalidAmountOrVersion'}
        '503': {$ref: '#/components/responses/BalanceSyncFailed'}
  /v1/card-balance-accounts/{cardBalanceAccountId}/wishes/{wishId}/withdrawals:
    parameters:
      - $ref: '#/components/parameters/CardBalanceAccountId'
      - $ref: '#/components/parameters/WishId'
    post:
      tags: [Wishes]
      operationId: withdrawFromWish
      summary: Withdraw funds from one Wish
      security: [{SeedBearer: []}]
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/WishAmountCommand'}
      responses:
        '200': {$ref: '#/components/responses/WishMutationSuccess'}
        '400': {$ref: '#/components/responses/MalformedOrIdempotencyRequired'}
        '401': {$ref: '#/components/responses/AuthRequired'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/WishOrAccountNotFound'}
        '409': {$ref: '#/components/responses/WithdrawalConflict'}
        '422': {$ref: '#/components/responses/InvalidAmountOrVersion'}
  /v1/card-balance-accounts/{cardBalanceAccountId}/transfers:
    parameters: [{$ref: '#/components/parameters/CardBalanceAccountId'}]
    post:
      tags: [Wishes]
      operationId: transferWishFunds
      summary: Atomically transfer funds between two Wishes in one account
      security: [{SeedBearer: []}]
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/WishTransferRequest'}
      responses:
        '200':
          description: Atomic transfer completed; identical replay returns the original status and body.
          headers:
            Idempotency-Replayed: {$ref: '#/components/headers/IdempotencyReplayed'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/WishTransferResult'}
        '400': {$ref: '#/components/responses/MalformedOrIdempotencyRequired'}
        '401': {$ref: '#/components/responses/AuthRequired'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/WishOrAccountNotFound'}
        '409': {$ref: '#/components/responses/TransferConflict'}
        '422': {$ref: '#/components/responses/InvalidTransferAmountOrVersion'}
...
components:
  securitySchemes:
    SeedBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque-seed-token
      description: Opaque deterministic principal token. A known Seed token identifies either a student or an authenticated non-student staff principal. Token issuance and refresh are outside this contract.
  parameters:
    CardBalanceAccountId:
      name: cardBalanceAccountId
      in: path
      required: true
      schema: {$ref: '#/components/schemas/Uuid'}
    WishId:
      name: wishId
      in: path
      required: true
      schema: {$ref: '#/components/schemas/Uuid'}
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Permanent per-student namespace; reuse is valid only for the same operation, target, and canonical request.
      schema: {type: string, minLength: 1, maxLength: 200}
...
  headers:
    IdempotencyReplayed:
      description: True only when the original status and body are replayed for an identical request.
      schema: {type: boolean}
  responses:
    WishMutationSuccess:
      description: Wish mutation completed; identical replay returns the original status and body.
      headers:
        Idempotency-Replayed: {$ref: '#/components/headers/IdempotencyReplayed'}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/WishMutationResult'}
...
    MalformedOrIdempotencyRequired:
      description: MALFORMED_REQUEST, including a missing or non-integer required version, or IDEMPOTENCY_KEY_REQUIRED.
      x-error-codes: [MALFORMED_REQUEST, IDEMPOTENCY_KEY_REQUIRED]
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
...
    AuthRequired:
      description: AUTH_REQUIRED — missing or invalid bearer token.
      x-error-codes: [AUTH_REQUIRED]
      headers:
        WWW-Authenticate:
          required: true
          schema: {type: string, const: Bearer}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
    Forbidden:
      description: FORBIDDEN — the authenticated principal is not a student.
      x-error-codes: [FORBIDDEN]
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
...
    WishOrAccountNotFound:
      description: CARD_BALANCE_ACCOUNT_NOT_FOUND or WISH_NOT_FOUND — absence, non-ownership, deletion, or hidden state.
      x-error-codes: [CARD_BALANCE_ACCOUNT_NOT_FOUND, WISH_NOT_FOUND]
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
...
    DepositConflict:
      description: VERSION_CONFLICT, INVALID_STATE_TRANSITION, BALANCE_MISMATCH_LOCKED, INSUFFICIENT_AVAILABLE_BALANCE, TARGET_AMOUNT_EXCEEDED, or IDEMPOTENCY_KEY_REUSED.
      x-error-codes: [VERSION_CONFLICT, INVALID_STATE_TRANSITION, BALANCE_MISMATCH_LOCKED, INSUFFICIENT_AVAILABLE_BALANCE, TARGET_AMOUNT_EXCEEDED, IDEMPOTENCY_KEY_REUSED]
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
          examples:
            balance-mismatch-locked: {$ref: '#/components/examples/BalanceMismatchLocked'}
    WithdrawalConflict:
      description: VERSION_CONFLICT, INVALID_STATE_TRANSITION, INSUFFICIENT_WISH_AMOUNT, or IDEMPOTENCY_KEY_REUSED.
      x-error-codes: [VERSION_CONFLICT, INVALID_STATE_TRANSITION, INSUFFICIENT_WISH_AMOUNT, IDEMPOTENCY_KEY_REUSED]
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
    TransferConflict:
      description: VERSION_CONFLICT, INVALID_STATE_TRANSITION, CROSS_ACCOUNT_TRANSFER_FORBIDDEN, INSUFFICIENT_WISH_AMOUNT, TARGET_AMOUNT_EXCEEDED, BALANCE_MISMATCH_LOCKED, or IDEMPOTENCY_KEY_REUSED.
      x-error-codes: [VERSION_CONFLICT, INVALID_STATE_TRANSITION, CROSS_ACCOUNT_TRANSFER_FORBIDDEN, INSUFFICIENT_WISH_AMOUNT, TARGET_AMOUNT_EXCEEDED, BALANCE_MISMATCH_LOCKED, IDEMPOTENCY_KEY_REUSED]
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
...
    InvalidAmountOrVersion:
      description: INVALID_AMOUNT or INVALID_VERSION — an independently decoded amount or expectedVersion violates its constraint.
      x-error-codes: [INVALID_AMOUNT, INVALID_VERSION]
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
          examples:
            invalid-expected-version: {$ref: '#/components/examples/InvalidExpectedVersion'}
    InvalidTransferAmountOrVersion:
      description: INVALID_AMOUNT or INVALID_VERSION — an independently decoded amount or source/destination version violates its constraint.
      x-error-codes: [INVALID_AMOUNT, INVALID_VERSION]
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
          examples:
            invalid-source-expected-version: {$ref: '#/components/examples/InvalidSourceExpectedVersion'}
            invalid-destination-expected-version: {$ref: '#/components/examples/InvalidDestinationExpectedVersion'}
...
    BalanceSyncFailed:
      description: BALANCE_SYNC_FAILED — retryable external balance query failure; the failed observation is persisted without mutating the Wish.
      x-error-codes: [BALANCE_SYNC_FAILED]
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
          examples:
            balance-sync-failed: {$ref: '#/components/examples/BalanceSyncFailed'}
  schemas:
    Uuid:
      type: string
      format: uuid
...
    KrwPositive:
      type: integer
      format: int64
      minimum: 1
      maximum: 9007199254740991
    WishVersion:
      type: integer
      format: int64
      minimum: 0
      description: A missing or non-integer required request version remains 400 MALFORMED_REQUEST; a decoded negative version returns 422 INVALID_VERSION; a stale non-negative version returns 409 VERSION_CONFLICT.
...
    WishAmountCommand:
      type: object
      additionalProperties: false
      required: [amount, expectedVersion]
      properties:
        amount: {$ref: '#/components/schemas/KrwPositive'}
        expectedVersion: {$ref: '#/components/schemas/WishVersion'}
...
    WishTransferRequest:
      type: object
      additionalProperties: false
      required: [sourceWishId, destinationWishId, amount, sourceExpectedVersion, destinationExpectedVersion]
      properties:
        sourceWishId: {$ref: '#/components/schemas/Uuid'}
        destinationWishId: {$ref: '#/components/schemas/Uuid'}
        amount: {$ref: '#/components/schemas/KrwPositive'}
        sourceExpectedVersion: {$ref: '#/components/schemas/WishVersion'}
        destinationExpectedVersion: {$ref: '#/components/schemas/WishVersion'}
    WishMutationResult:
      type: object
      additionalProperties: false
      required: [wish, eventId]
      properties:
        wish: {$ref: '#/components/schemas/Wish', description: "Authoritative Wish snapshot after the mutation, or the original snapshot returned by an identical idempotent replay."}
        eventId: {type: [string, 'null'], format: uuid, description: "UUID of the immutable ledger event created by the mutation; null when the mutation moves no funds and therefore creates no ledger event."}
    WishTransferResult:
      type: object
      additionalProperties: false
      required: [sourceWish, destinationWish, eventId, occurredAt]
      properties:
        sourceWish: {$ref: '#/components/schemas/Wish', description: "Authoritative source Wish snapshot after the atomic transfer."}
        destinationWish: {$ref: '#/components/schemas/Wish', description: "Authoritative destination Wish snapshot after the atomic transfer."}
        eventId: {$ref: '#/components/schemas/Uuid', description: "UUID of the single immutable ledger event containing both transfer effects."}
        occurredAt: {$ref: '#/components/schemas/UtcInstant', description: "RFC 3339 UTC Z instant shared by the source and destination effects of the transfer."}
...